Update and rename 11651.cpp to 11651_1.cpp
